requirements.txt
Django==4.2.11
django-filter==24.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-extra-fields==3.7.0
environs==11.0.0
pyTelegramBotAPI==4.16.1
Pillow==10.2.0
